refactor: ErrorMessage (형식, 비즈니스) 분리 리팩토링
