+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.view.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
